Create Guessing Game final version
The final version of the game with a guess counter added.  The game will now generate a random number and accept user input for guesses.  It will then compare the guess against the random number and display the appropriate message based in that guess.  It will also tell the user how many guesses they have done.